Aspect based emitter sinks! Coolio stuff (#12)
* [FEAT] Updated dependencies

* [FIX] Undone local changes from task definition

* [FEAT] Created Duplex implementation

* [FEAT] Added reflective duplex initialization

* [FEAT] Added missing duplex annotation

* [FEAT] Removed old emitter-sink model files

* [FEAT] Added logger for duplex initialization
