+++ build.gradle
@@ -30,33 +30,36 @@ compileKotlin {
 
 dependencies {
     // SLF4J
-    implementation 'org.apache.logging.log4j:log4j-slf4j-impl:2.14.1'
+    implementation 'org.apache.logging.log4j:log4j-slf4j-impl:2.16.0'
 
     // JUnit
-    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.8.1'
-    testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.8.1'
+    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.8.2'
+    testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.8.2'
 
     // JDA
-    implementation("net.dv8tion:JDA:4.3.0_333")
+    implementation("net.dv8tion:JDA:4.4.0_352")
 
     // Jackson XML
     implementation 'com.fasterxml.jackson.dataformat:jackson-dataformat-xml:2.13.0'
 
     // OkHttp4
-    implementation('com.squareup.okhttp3:okhttp:4.9.2')
+    implementation('com.squareup.okhttp3:okhttp:4.9.3')
 
     // Reactor
-    testImplementation 'io.projectreactor:reactor-test:3.4.10'
-    implementation 'io.projectreactor:reactor-core:3.4.10'
+    testImplementation 'io.projectreactor:reactor-test:3.4.13'
+    implementation 'io.projectreactor:reactor-core:3.4.13'
 
     // Annotations
     compileOnly 'org.jetbrains:annotations:22.0.0'
 
-    // Lavaplayer
+    // LavaPlayer
     implementation 'com.sedmelluq:lavaplayer:1.3.78'
 
     // Apache-Commons
     implementation 'org.apache.commons:commons-lang3:3.12.0'
+
+    // Reflections
+    implementation 'org.reflections:reflections:0.10.2'
 }
 
 ext {
