+++ src/main/java/personal/opensrcerer/aspect/PreDuplex.java
@@ -1,0 +1,14 @@
+package personal.opensrcerer.aspect;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+/**
+ * Initializes Duplexes before the JDA Ready Signal.
+ */
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.TYPE)
+public @interface PreDuplex {
+}
