+++ src/main/java/personal/opensrcerer/duplex/impl/slash/player/PauseDuplex.java
@@ -1,0 +1,20 @@
+package personal.opensrcerer.duplex.impl.slash.player;
+
+import net.dv8tion.jda.api.events.interaction.SlashCommandEvent;
+import personal.opensrcerer.aspect.PostDuplex;
+import personal.opensrcerer.config.SlashCommand;
+import personal.opensrcerer.duplex.abstractions.SlashCommandDuplex;
+import personal.opensrcerer.services.audio.MusicPlayer;
+
+@PostDuplex
+public class PauseDuplex extends SlashCommandDuplex {
+    public PauseDuplex() {
+        super(SlashCommand.PAUSE.getName());
+    }
+
+    @Override
+    public void onEvent(SlashCommandEvent event) {
+        MusicPlayer.MUSIC_PLAYER.pause(event.getGuild());
+        event.reply("Player paused!").queue();
+    }
+}
