+++ src/main/java/personal/opensrcerer/duplex/interfaces/Duplex.java
@@ -1,0 +1,6 @@
+package personal.opensrcerer.duplex.interfaces;
+
+import net.dv8tion.jda.api.events.GenericEvent;
+
+public interface Duplex<E extends GenericEvent> extends Emitter, Sink<E> {
+}
