+++ src/main/java/personal/opensrcerer/duplex/interfaces/Emitter.java
@@ -1,0 +1,9 @@
+package personal.opensrcerer.duplex.interfaces;
+
+@FunctionalInterface
+public interface Emitter {
+    /**
+     * Starts emitting events.
+     */
+    void emit();
+}
