+++ src/main/java/personal/opensrcerer/launch/SupersonicLaunchWrapper.java
@@ -4,7 +4,7 @@ import net.dv8tion.jda.api.JDABuilder;
 import net.dv8tion.jda.api.entities.Activity;
 import net.dv8tion.jda.api.utils.MemberCachePolicy;
 import net.dv8tion.jda.api.utils.cache.CacheFlag;
-import personal.opensrcerer.reactive.emitters.system.ReadyEmitter;
+import personal.opensrcerer.aspect.DuplexInitializer;
 
 import javax.security.auth.login.LoginException;
 import java.util.stream.Collectors;
@@ -30,6 +30,6 @@ public abstract class SupersonicLaunchWrapper {
                         )
                 ).build();
 
-        new ReadyEmitter().emit();
+        DuplexInitializer.initializePreDuplexes();
     }
 }
