--- src/main/java/personal/opensrcerer/reactive/emitters/Emitter.java
@@ -1,12 +1,0 @@
-package personal.opensrcerer.reactive.emitters;
-
-/**
- * Groups all Event Emitters together.
- */
-@FunctionalInterface
-public interface Emitter {
-    /**
-     * Subscribes to a sink and emits events when received.
-     */
-    void emit();
-}
