--- src/main/java/personal/opensrcerer/reactive/emitters/slash/player/UnpauseEmitter.java
@@ -1,14 +1,0 @@
-package personal.opensrcerer.reactive.emitters.slash.player;
-
-import personal.opensrcerer.config.SlashCommand;
-import personal.opensrcerer.reactive.emitters.slash.SlashEmitter;
-import personal.opensrcerer.reactive.sinks.slash.player.UnpauseSink;
-
-public class UnpauseEmitter extends SlashEmitter {
-    public UnpauseEmitter() {
-        super(
-                SlashCommand.UNPAUSE.getName(),
-                new UnpauseSink()
-        );
-    }
-}
