--- src/main/java/personal/opensrcerer/reactive/emitters/system/ReadyEmitter.java
@@ -1,15 +1,0 @@
-package personal.opensrcerer.reactive.emitters.system;
-
-import net.dv8tion.jda.api.events.ReadyEvent;
-import personal.opensrcerer.reactive.sinks.system.ReadySink;
-import personal.opensrcerer.reactive.emitters.DiscordEmitter;
-
-public class ReadyEmitter extends DiscordEmitter<ReadyEvent> {
-
-    public ReadyEmitter() {
-        super(
-                ReadyEvent.class,
-                new ReadySink()
-        );
-    }
-}
