--- src/main/java/personal/opensrcerer/reactive/sinks/components/ButtonClickSink.java
@@ -1,12 +1,0 @@
-package personal.opensrcerer.reactive.sinks.components;
-
-import net.dv8tion.jda.api.events.interaction.ButtonClickEvent;
-import personal.opensrcerer.reactive.sinks.Sink;
-import personal.opensrcerer.services.pagination.PaginationService;
-
-public class ButtonClickSink implements Sink<ButtonClickEvent> {
-    @Override
-    public void onEvent(ButtonClickEvent event) {
-        PaginationService.decode(event);
-    }
-}
