--- src/main/java/personal/opensrcerer/reactive/sinks/slash/player/PauseSink.java
@@ -1,19 +1,0 @@
-package personal.opensrcerer.reactive.sinks.slash.player;
-
-import net.dv8tion.jda.api.Permission;
-import net.dv8tion.jda.api.events.interaction.SlashCommandEvent;
-import personal.opensrcerer.services.audio.MusicPlayer;
-import personal.opensrcerer.reactive.sinks.slash.SlashCommandSink;
-
-public class PauseSink extends SlashCommandSink {
-
-    public PauseSink(Permission... permissions) {
-        super(permissions);
-    }
-
-    @Override
-    public void onEvent(SlashCommandEvent event) {
-        MusicPlayer.MUSIC_PLAYER.pause(event.getGuild());
-        event.reply("Player paused!").queue();
-    }
-}
