--- src/main/java/personal/opensrcerer/reactive/sinks/slash/player/UnpauseSink.java
@@ -1,19 +1,0 @@
-package personal.opensrcerer.reactive.sinks.slash.player;
-
-import net.dv8tion.jda.api.Permission;
-import net.dv8tion.jda.api.events.interaction.SlashCommandEvent;
-import personal.opensrcerer.services.audio.MusicPlayer;
-import personal.opensrcerer.reactive.sinks.slash.SlashCommandSink;
-
-public class UnpauseSink extends SlashCommandSink {
-
-    public UnpauseSink(Permission... permissions) {
-        super(permissions);
-    }
-
-    @Override
-    public void onEvent(SlashCommandEvent event) {
-        MusicPlayer.MUSIC_PLAYER.unpause(event.getGuild());
-        event.reply("Player unpaused!").queue();
-    }
-}
